Bugfixes + Killicons
Bugfixes:
Fixed an issue with scoped base weapons not resetting the mouse sensitivity correctly when holstering a weapon while zoomed in.
Fixed an issue with the Thompson Contender G2 not unzooming when playing any of the two reload animations.
Fixed an issue with the Matador not unzooming when reloading.
Fixed an issue with the M202 still being able to zoom in while reloading.
Fixed an issue with the EX41 starting with 3 grenades in the clip. Only rocket launchers are supposed to start with ammo in the "magazine".

Re-added Killicons code. They should work again, though I don't even know if they were broken in the first place.

Optimizations:
Removed pointless function in the gun base.