QoL Improvement for Ironsights + Bugfix
Ironsights have received a quality of life improvement.
We are sacrificing a bit of detail by forcing the weapon to play a simplified animation while aiming down the sights. This makes it A LOT easier to see things

Fixed problematic CanPrimaryAttack statements.
Fixed Shotty based weapons not having a MuzzleFlash light effect.