Bugfix
Bugfixes (Singleplayer)
Fixed a bug when 'opening' an ammo crate and receiving the weapon the game would make you select the weapon but not actually do it. Sort of like a ghost state. This was fixed by explicitly telling the game to make you select that weapon. (This does nothing in multiplayer)