Begone ironsight viewmodel sway! + Bugfixes
New Features:
Drastic changes to the viewmodel swaying have been made to make it so that there is barely any sway while aiming down the sights and thus ironsights have become WAY more useable!
Changed the way spread through moving is handled. The new method allows modders and pack creators to assign a variable to their weapon class to disable this feature if desired

Bugfixes:
Fixed the absurd ironsight of the AAC Honey Badger. The player now holds the gun a tiny bit further away from their face so the lever doesn't stick in their face!
Fixed muzzleflash effect playing even when scoped in
Fixed a bug with shotguns when using ironsights and forcing a reload through CanPrimaryAttack (Trying to fire while magazine is empty), going out of prediction and causing the ironsights to be stuck on the client until they are re-used
Fixed duplicate $basematerial error regarding t_phoenix material
Fixed an error that would occur when a weapon is missing a required attachment