Re-added Scopes + Bugfixes
Features:
Previously a few scopes have been removed from weapons as they were rendered with bad code. With the recently newly added grenade base the techology to do it efficiently exists so I've gone ahead and re-added them.
- The AAC Honey Badger has a scope again and can make use of it with improved accuracy while scoping
- The KAC PDW has a scope again
- The Milkor Mk1 has a scope again
- The FN P90 has a scope now. The Legacy M9k version did not have one

Bugfixes:
Fixed a bug with bolt-action sniper rifles restoring the zoom while reloading. This happened when reloading the sniper rifle while the player was cycling a new bullet
Weapons using the ACOG or Aimpoint scopes can no longer zoom in multiple times. These scopes do not have a zoom functionality. These weapons include:
- AUG A3
- AAC Honey Badger (New)
- M16A4 ACOG
- L85
- L85 (Admin)
- F2000
- SL8
- MP7

Optimizations:
Removed SpreadZoomed variable from the Thompson Contender G2 as it was the exact same as in the base
Removed 'OurClass' definitions and replaced them with self.ClassName as it is exactly the same